estrutura basica (ainda aprendendo pygame)
